// src/Types/search/search.ts
export enum TypesSearch {
   SEARCH_NEWS = "news",
   SEARCH_SALES = "sales",
   SEARCH_WEAR = "no_shoes",
   SEARCH_SHOES = "shoes",
   SEARCH_LUXERY = "luxery_brand",
   SEARCH_MODA = "moda",
   SEARCH_ALL = "all",

   SEARCH_TSHIRT = 't-shirt',
   SEARCH_BOTTOM = 'bottom',

   SEARCH_MARTINS = 'dr.martins',
   SEARCH_TNF = 'the_north_face',
   SEARCH_NIKE = 'nike',
   SEARCH_CARHARTT = 'carhartt',
   SEARCH_ELLESSE = 'ellesse'
}